// FinancialPortfolioClient/FinancialPortfolio.Client/src/core/models/portfolio/portfolio-position.model.ts
export interface PortfolioPosition {
  serialNo: number;
  stockId: number;
  symbol: string;
  companyName: string;
  logoUrl?: string | null;
  stockCode: string;
  exchange: number;
  currentQuantity: number;
  lifetimeBoughtQuantity: number;
  lifetimeSoldQuantity: number;
  averageBuyPrice: number;
  lifetimeAverageBuyPrice: number;
  averageSellPrice?: number | null;
  marketPrice: number;
  totalInvestment: number;
  lifetimeInvestment: number;
  totalCurrentValue: number;
  totalOnSell: number;
  unrealizedGainLoss: number;
  unrealizedGainLossPercent: number;
  realizedGainLoss: number;
  realizedGainLossPercent: number;
  totalDividends: number;
  totalGainLoss: number;
  gainLossPercent: number;
  profitLoss: number;
  holdDays: number;
  status: number;
  currentType: number;
  buyLotCount: number;
  openLotCount: number;
  sellCount: number;
  dividendCount: number;
  firstPurchaseDate: string;
  lastExitDate?: string | null;
  lastActivityDate: string;
  asOfDate: string;
}